==> DNNREC_FilmTrust/1. Data Preparation.ipynb <==
PATH='/home/kirana/Documents/phd/experiment1_new/DNNREC_FilmTrust'

DATAPATH='/home/kirana/Documents/final_dissertation_final/experiments/datasets/filmtrust'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sklearn
import os
import pickle
import torch
from torch.utils.data import Dataset, DataLoader
from torch import nn
import pickle
from sklearn import metrics

ratings=pd.read_csv(f'{DATAPATH}/ratings.txt',sep=' ',names=['userId','itemId','rating'])
print (ratings.head())

ratings.shape

ratings.drop_duplicates(inplace=True)

ratings[['userId','itemId']].duplicated().value_counts()

ratings=ratings.groupby(['userId','itemId'],as_index=False)['rating'].mean()

ratings.shape

trust=pd.read_csv(f'{DATAPATH}/trust.txt',sep=' ',names=['userId','userId_1','trust'])
print (trust.head())

ratings.shape, trust.shape

trust_agg=trust.groupby('userId')['trust'].agg(['mean','min','max','sum'])

trust_agg.columns+='_on_other'

trust_agg.head()

trust_agg_on_self=trust.groupby('userId_1')['trust'].agg(['mean','min','max','sum'])

trust_agg_on_self.columns+='_on_self'

trust_agg_on_self.head()

trust_agg_on_self.reset_index(inplace=True)

trust_agg.reset_index(inplace=True)

trust_agg_on_self.rename({'userId_1':'userId'},axis=1,inplace=True)

trust_agg_on_self.head()

ratings.shape

trust_agg_on_self.columns

trust_agg.columns

trust_agg.columns='f_'+trust_agg.columns

trust_agg.rename({'f_userId':'userId'},axis=1,inplace=True)

trust_agg.columns

trust_agg_on_self.columns='f_'+trust_agg_on_self.columns

trust_agg_on_self.rename({'f_userId':'userId'},axis=1,inplace=True)

trust_agg_on_self.columns

print(ratings.shape)
ratings=pd.merge(ratings,trust_agg_on_self,on='userId',how='left')
print(ratings.shape)

print(ratings.shape)
ratings=pd.merge(ratings,trust_agg,on='userId',how='left')
print(ratings.shape)

ratings.tail()

ratings.head()

ratings.shape

dfratings=ratings.copy()

ratings.head()

### CROSS-VALIDATION - SIMULATE COLD START

from sklearn.model_selection import train_test_split

ratings['rating'].value_counts()

ratings.head()

dfratings=ratings.copy()

users=dfratings['userId'].unique()

users=np.random.permutation(users)

normalfrac=0.9

users_normal=users[:np.int(np.round(len(users)*normalfrac))]

dfnormal=dfratings.loc[dfratings['userId'].isin(users_normal)]

dfsimulate=dfratings.loc[~dfratings['userId'].isin(users_normal)]

dfnormal.shape, dfsimulate.shape

temptrain,tempvalid=train_test_split(dfnormal,test_size=0.1,train_size=0.9,random_state=11,shuffle=True)

temptrain['random_dstype']='train'
tempvalid['random_dstype']='test'

dfnormal=pd.concat([temptrain,tempvalid],axis=0)

dfnormal.shape

dfnormal['random_dstype'].value_counts()

dfsimulate.shape

simulateusers=dfsimulate['userId'].unique()

simulateusers=np.random.permutation(simulateusers)

simulateusers_train=simulateusers[:np.int(np.round(len(simulateusers)*normalfrac))]

dfsimulate_train=dfsimulate.loc[dfsimulate['userId'].isin(simulateusers_train)]
dfsimulate_test=dfsimulate.loc[~dfsimulate['userId'].isin(simulateusers_train)]

dfsimulate_train['random_dstype']='train'
dfsimulate_test['random_dstype']='test'

dfsimulate=pd.concat([dfsimulate_train,dfsimulate_test],axis=0)

dfratings=pd.concat([dfnormal,dfsimulate],axis=0)

dftrain=dfratings.loc[dfratings['random_dstype']=='train']

dfvalid=dfratings.loc[dfratings['random_dstype']!='train']

dftrain.shape, dfvalid.shape

dfvalid['random_dstype'].value_counts()

dftrain['flag_train']=1
dfvalid['flag_train']=0

### Missing value treatment

user_features=dftrain.columns[dftrain.columns.str.startswith('f_')]

user_features

dftrain[user_features].describe().T

for myNm in user_features:
    dftrain['f_na_'+myNm[2:]]=dftrain[myNm].isnull()
    dftrain[myNm].fillna(0,inplace=True)

dftrain[user_features].describe().T

for myNm in user_features:
    dfvalid['f_na_'+myNm[2:]]=dfvalid[myNm].isnull()
    dfvalid[myNm].fillna(0,inplace=True)

pickle.dump([dftrain,dfvalid],open(f'{DATAPATH}/df.pkl','wb'))
